combine Primitive tests
